Add server status retrieval (#7522)
## Product change and motivation
Add server status APIs to `ServerState` to retrieve server addresses and
replica statuses. The general goal is to be able to share these states
across replicas and return them to the clients requesting `servers_all`.

## Implementation
Rename `ServerInfo` to `DistributionInfo`.
Introduce `ServerStatus` trait with a `LocalServerStatus`
implementation, containing server addresses as strings (unparsed -- we
will lose the parsed variant across multiple protocol layers anyway).
This trait will be implemented in Cluster as `ClusteredServerStatus`, an
extended version with replication information.
Extend `ServerState` to return `server_status` for a single server and
`servers_statuses` for multiple servers as a response to `servers_all`
grpc request.