Lock attributes on updating connections (#7566)
## Product change and motivation

We fix an Isolation bug that is exposed under concurrent update (adding
an ownership) + concurrent delete transactions. This is a relatively
uncommon conflict, and is fixed by locking attributes that are edited in
order to conflict with concurrent deletes of the same attribute. Under
highly concurrent operations that include `delete`s, this change might
manifest itself as more transaction conflicts `STC2` (Storage Commit 2)
errors, which can be resolved with a retry.

## Implementation

We mark attributes that are modified (eg. given new ownerships) as
required by locking them, rather than re-putting them into the snapshot.

We also fix a similar issue with links edges, by marking the player and
the relation as locked so a concurrent delete of either will not lead a
dangling edge.

## Details

Scenario: attribute A exists beforehand with 1 owner X

Tx1: opens at SeqNr 2
  --> deletes last ownership it sees from X to A
  --> commits
  --> auto-delete dependent attribute A
  --> commits, gets sequence number 8

Tx2: opens at SeqNr 4
  --> adds ownership of A from owner Y, and re-puts A
  --> commits
  --> commits at seq nr 6
  --> NO ERROR, and commits first !

Tx1: NO ERROR. Even though we delete attribute A, we only conflict
"locks" and "deletes". We didn't set attributes as "locked" when we add
attribute ownerships to them, but we do re-insert them - but this isn't
enough!