Only seek iterators in k-way merge when behind (#7580)
## Product change and motivation

Fix intermittent crashes with `Key behind the stored item in a Peekable
iterator` when a roleplayer index constraint is part of a join.

## Implementation

In general, it is considered an error to seek an iterator that is
already ahead of the seek target. We cannot seek backwards, and we
cannot guarantee that the iterator is pointing at the element one after
the target, which is what the caller code expects.

In a `Peekable<I>`, if the peeked item is behind the target, but the tip
of the iterator `I` is ahead _and_ the `I` can detect that (being
another `Peekable` at some level), we run into an issue where this
condition is checked multiple times, causing a crash.

We relax this constraint to only apply to `Peekable`. When `seek` is
called, it peeks the current item in the iterator (panicking if `seek`
was used incorrectly to avoid illegal state), then, if needed, calls
`seek` on the inner iterator which advances unconditionally.

Closes #7577, closes #7446.