Make state initialisation and loading extensible (#7587)
The state initialisation and loading have been made extensible as needed
by TypeDB Cluster.

- Refactor state initialise / load methods in order to make them
extensible
- Make user creation method in user management return a commit record
for extensibility. This way, the commit record can be handled either by
local or cluster-wide database commit procedure
- Do database management operations through the "unrestricted" method
counterparts, so that they can be used for performing system database
initialisation
- Fix a bug where the 'load' procedure is never triggered. Instead what
had always happened was that the initialise procedure is always called.
It happened to work since the initialise method is idempotent.
- Fix a bug where the system database isn't loaded in the load procedure

---------

Co-authored-by: Georgii Novoselov <georgii@typedb.com>