Reenable Sentry crash reporting (#7590)
## Product change and motivation
Fix a bug when TypeDB Server did not send crash reports for diagnostics
even when `--diagnostics.reporting.errors` was enabled.

## Implementation
Fixes the bug introduced in https://github.com/typedb/typedb/pull/7391.

Preserve the ownership of the Sentry's `Guard` until the `main` function
is exited. To avoid the same mistake, we encapsulate all the guards and
other things required to remain alive while the App is running inside a
new struct called `ServerApplication`.