Add WAL truncation methods  (#7560)
## Product change and motivation
Expose public interfaces to request WAL's truncation starting from a
sequential number. This is a required feature to use WAL as a serial
storage for TypeDB Cluster's Raft log, which allows and expects record
overriding.

Additionally, enhance storage checkpoints' correctness guarantees,
minimising the risk of reading invalid data.

## Implementation
Add `truncate_from(seqnum)` to `DurabilityService` and its
implementation to `WAL`. This methods removes all the records (and, if
needed, emptied files) starting from the provided sequence number. Add
related unit tests.

Modify `recovery/checkpoint.rs`, providing safer read and write methods
to work with checkpoint files. Instead of directly writing to the target
file, write to a temporary file, fsync the temporary file, and then
rename it to the final state. This makes sure that the target file, if
it exists, is complete, and its content is 100% safe. However, if there
was a crash before the final (atomic) rename, the temporary file is
ignored, and the recovery process reads the previous file instead.