Print ready message only for the same distribution (#7610)
## Product change and motivation
Restrict the "Ready!" message printed on the bootup to only CE
distributions of TypeDB. This allows other distributions with
potentially more complicated initialization to handle the readiness
notifications themselves.

## Implementation
Comparing `DISTRIBUTION` directly is the easiest path. Introducing
another configurable boolean flag is too much for this job. Requiring
changing the distribution name for such controls must be a decent
tradeoff for such a minor piece, although it's not very clean.