Ensure graceful shutdown for interval tasks (#7623)
## Product change and motivation
Add a tracker for `tokio` tasks spawned in the server's components that
require graceful shutdown when the server closes. `TokioIntervalRunner`,
previously used for such tasks, used to block the `tokio` runtime, which
worked for an isolated version of the CE server, but [led to
crashes](https://tokio.rs/tokio/topics/shutdown) in more complex
applications reusing TypeDB (e.g., TypeDB Cluster).

## Implementation

* Remove `TokioIntervalRunner`;
* Add `TokioTaskTracker` used in the server to track and join these
background tasks;
* Add `TokioTaskSpawner` as the clonable part of the `TokioTaskTracker`,
which is sent to `TokenManager`, diagnostics `Reporter`, and HTTP
`TransactionService`;
* Cover these new components with unit tests;
* Instead of saving the task handles in the components, just spawn the
tasks from `TokioTaskSpawner`s and rely on the `TokioTaskTracker` to
manage them;
* Await all tracked tasks after TypeDB services are closed.

We currently track only interval tasks because they're the only tasks
that require a graceful shutdown. Other tokio tasks can be spawned
untracked, as usual.