Add in panic logging via the global panic_hook (#7628)
## Product change and motivation

We ensure that panics are written to the configured `tracing` log file,
by intercepting the panic event and routing it via the logger. This code
is borrowed from https://github.com/LukeMathWalker/tracing-panic !


## Implementation

We decided not to use the published crate for such a small function to
minimize dependencies and future security validation work.