Improve type-seeding failure error experience  (#7630)
## Product change and motivation
We improve the error message returned when type-seeding fails before the iterative pruning step by including the constraint name in the message.
We also make type-inference propagate labels across 'isa' constraints first, since these are likely to be the most informative. This makes for more intuitive error messages when the 'isa' constraint is the unsatisfiable one.