Generalize database import and clean up transactions (#7640)
Move database import to `ServerState` to override/disable it in TypeDB
Cluster. Clean up transaction-related utils and TODOs.

We want to temporarily disable database import in TypeDB Cluster because
of the complications of all its stages' replication (e.g., the cleanup
after an unsuccessful import attempt). To do so, we put the main task of
the importer into a server state's method.