Reduce diagnostics noise from CI of TypeDB users (#7651)
## Product change and motivation

Add a requirement for `initial_delay` of diagnostics reports, which, if
not met, forces the reporter to skip the first report cycle. This allows
silencing CI jobs built by the users of TypeDB not using the
`diagnostics.reporting.metrics` flag without significant harm to the
data (the data will be added to the next report unless the server stops
early).

Additionally, clean up the diagnostics logic from the outdated code from
TypeDB 2.x not used in TypeDB 3.x.

## Implementation

* Update `reporter.rs`: add `REPORT_INTERVAL` to the calculated initial
delay if it's too small
* Remove `is_owned` flag of the database metrics: all databases are
"owned" by each replica of Cluster. In 3.x, the primary replica is
selected for the whole TypeDB infrastructure, not per database. Thus,
this flag is not needed.