Ignore already processed commit records in storage (#7638)
Add locally generated `snapshot_id`s for snapshots and `CommitRecord`s.
These ids are used in combination with `open_sequence_number` to
uniquely identify currently opened snapshots and transactions, as well
as committed commit records.

A new method `commit_record_exists` is added. It can be used to check if
a `CommitRecord` with the provided `open_sequence_number` and
`snapshot_id` exists locally. This method can be useful for an extension
of TypeDB. E.g., TypeDB Cluster uses this method to avoid replaying
already existing `CommitRecord`s on recovery, protecting idempotency.

**Important**: Currently, snapshot ids are generated locally, which can
cause collisions if you depend on them without enough awareness.
**However**, they are totally safe for a single instance of TypeDB.

**Also important**: `commit_record_exists` can be quite expensive if the
record is old, and can be **wrong** if the record is **very** old (thus,
removed from the WAL). It's recommended to use it in a limited way.

This change **does not** break backward compatibility. Databases created
using prior versions of TypeDB will successfully be loaded. However, you
will not be able to use a database modified in this version by an older
release of TypeDB.

Add new methods to serialize transaction openings and manually close
specific transactions through TypeDB Cluster. This helps TypeDB Cluster
make its logic safe and friendly to the user, only providing up-to-date
data, and protects the uniqueness of snapshot ids in a distributed
environment.

Add `snapshot_id` to `CommitRecord`, extending the structure by 9 bytes
(`u64`). Additionally, save the same value in Snapshots.

In `storage`, add a method to check if a commit record with the same
seqnum and id exists.

To be able to close transactions without a user command, we add a new
channel between transaction services and `ServerState`. The
`ServerState` is able to close transactions based on their type on
request. This mechanism will also allow us to implement transaction
invalidation on the owner's credential updates.

For other implementation details, check out the PR comments.